Filter Features and Variables with High Missingness

Removes exposure variables and omics features with missing values above a specified threshold. Generates missing data summaries and quality control (QC) plots.

Usage

filter_missing(expomicset, na_thresh = 20, na_plot_thresh = 5)

Arguments

expomicset

A MultiAssayExperiment object containing exposure and omics data.

na_thresh

A numeric value specifying the percentage of missing data allowed before a variable or feature is removed. Default is 20.

na_plot_thresh

A numeric value specifying the minimum missing percentage for inclusion in QC plots. Default is 5.

Value

A MultiAssayExperiment object with filtered exposure variables and omics features. QC results, including missingness summaries and plots, are stored in metadata(expomicset)$na_qc.

Details

The function assesses missingness in both colData(expomicset) (exposure data) and experiments(expomicset) (omics data).

Exposure variables with more than na_thresh% missing values are removed.

Omics features (rows in assay matrices) exceeding na_thresh% missing values are filtered.

Missingness summaries and QC plots are generated using naniar::gg_miss_var() and stored in metadata.

Examples

if (FALSE) { # \dontrun{
filtered_expom <- filter_missing(
  expomicset = expom,
  na_thresh = 20,
  na_plot_thresh = 5
)
} # }