Sample-Level Metadata for the Exposome Challenge Dataset

A data.frame of maternal-child metadata including sample identifiers, demographic information, clinical outcomes and exposure information.

Usage

meta

Format

A data.frame with one row per sample and columns including:

hs_child_age_None

Child age in years.

e3_sex_None

Biological sex (Male/Female).

h_cohort

Cohort identifier.

hs_asthma

Binary indicator of asthma outcome.

...

Additional exposure and phenotype variables.

Source

Derived from the ISGlobal Exposome Data Challenge 2021.

Examples

data(meta)
head(meta)
#>       h_abs_ratio_preg_Log h_no2_ratio_preg_Log h_pm10_ratio_preg_None
#> s548            0.85260563             3.373861               22.44994
#> s555            0.71601007             3.024288               21.90104
#> s598           -0.35482399             3.373861                9.01958
#> s395            0.29767305             2.864948               13.98486
#> s1135           0.57283889             2.541408               37.36215
#> s178            0.09775766             2.472750               13.58145
#>       h_pm25_ratio_preg_None hs_no2_dy_hs_h_Log hs_no2_wk_hs_h_Log
#> s548               18.016866           3.658223           3.571212
#> s555               18.961571           2.501568           2.638058
#> s598                8.954145           1.221778           1.818556
#> s395                9.339700           2.567528           2.798095
#> s1135              13.853442           2.369783           2.206939
#> s178               14.879898           1.104555           1.230386
#>       hs_no2_yr_hs_h_Log hs_pm10_dy_hs_h_None hs_pm10_wk_hs_h_None
#> s548            3.229627            37.029556            34.673076
#> s555            2.688654            13.646372            29.141274
#> s598            1.617848             9.771050            13.421016
#> s395            3.067854             8.938537             6.503934
#> s1135           2.296243            36.456271            36.137574
#> s178            1.940459            29.817442            23.862842
#>       hs_pm10_yr_hs_h_None hs_pm25_dy_hs_h_None hs_pm25_wk_hs_h_None
#> s548              24.12191            20.442523            16.524717
#> s555              39.06746            13.704656            22.524238
#> s598              11.68626             5.902195             6.317749
#> s395              12.69163             2.433092             3.771206
#> s1135             33.87148            13.835949            11.756729
#> s178              23.96526             4.446991             5.043293
#>       hs_pm25_yr_hs_h_None hs_pm25abs_dy_hs_h_Log hs_pm25abs_wk_hs_h_Log
#> s548             10.325866              0.9413520              0.7506622
#> s555             16.524127             -0.1912293              0.2311378
#> s598              4.829015             -1.1806855              0.2503305
#> s395              7.650365             -0.8855793             -0.3277950
#> s1135            12.978987              0.9155942             -0.1380054
#> s178             10.258761              0.1382814             -0.1380054
#>       hs_pm25abs_yr_hs_h_Log h_accesslines300_preg_dic0
#> s548               0.3103651                          0
#> s555               0.1739284                          0
#> s598              -0.2512695                          0
#> s395              -0.2353678                          0
#> s1135              0.1736764                          0
#> s178               0.2510297                          0
#>       h_accesspoints300_preg_Log h_builtdens300_preg_Sqrt
#> s548                    3.667571                352.46231
#> s555                    2.656014                362.81048
#> s598                    3.667571                493.82564
#> s395                    2.879144                251.47892
#> s1135                   1.269931                637.87320
#> s178                    3.466906                 82.64075
#>       h_connind300_preg_Sqrt h_fdensity300_preg_Log h_frichness300_preg_None
#> s548               10.160145               12.46006               0.05263158
#> s555                4.218770               10.26284               0.00000000
#> s598               17.394436               11.87227               0.12280702
#> s395               10.838217               11.36145               0.01754386
#> s1135              18.961015               10.26284               0.00000000
#> s178                3.773383               10.26284               0.00000000
#>       h_landuseshan300_preg_None h_popdens_preg_Sqrt
#> s548                   0.8122130            74.98000
#> s555                   0.4314893            84.98823
#> s598                   0.4532843            74.98000
#> s395                   0.4645479            68.86786
#> s1135                  0.3904512            97.75480
#> s178                   0.2546103            41.54516
#>       h_walkability_mean_preg_None hs_accesslines300_h_dic0
#> s548                         0.350                        0
#> s555                         0.200                        0
#> s598                         0.350                        0
#> s395                         0.200                        0
#> s1135                        0.275                        0
#> s178                         0.150                        0
#>       hs_accesspoints300_h_Log hs_builtdens300_h_Sqrt hs_connind300_h_Log
#> s548                 2.1860528               523.2890            4.958529
#> s555                 2.5224930               331.8434            3.834599
#> s598                 3.5209711               133.3942            5.220893
#> s395                 3.6210518               250.3050            4.766157
#> s1135                0.5770643               352.8191            5.119797
#> s178                 3.5209711               330.9970            4.214089
#>       hs_fdensity300_h_Log hs_landuseshan300_h_None hs_popdens_h_Sqrt
#> s548              10.26284                0.5664721         83.466400
#> s555              10.26284                0.5472160         10.246951
#> s598              11.64913                0.3727202          5.140307
#> s395              10.95598                0.3741057         13.228757
#> s1135             10.26284                0.3093658          6.244998
#> s178              10.26284                0.5446557         33.778692
#>       hs_walkability_mean_h_None hs_accesslines300_s_dic0
#> s548                       0.375                        1
#> s555                       0.275                        0
#> s598                       0.275                        0
#> s395                       0.250                        0
#> s1135                      0.300                        0
#> s178                       0.200                        0
#>       hs_accesspoints300_s_Log hs_builtdens300_s_Sqrt hs_connind300_s_Log
#> s548                 2.5224930               402.8545            4.636945
#> s555                 3.4097490               503.4886            4.636945
#> s598                 1.6753021               127.0561            4.042238
#> s395                 1.6753021               221.4176            4.703637
#> s1135                0.5770643               300.8570            4.527746
#> s178                 3.2845902               404.2621            4.796010
#>       hs_fdensity300_s_Log hs_landuseshan300_s_None hs_popdens_s_Sqrt
#> s548              11.64913                0.5284814         69.469651
#> s555              12.05459                0.5515236          6.480741
#> s598              13.72857                0.3310432          3.328225
#> s395              10.95598                0.5713983         48.756297
#> s1135             10.26284                0.4754483         48.948953
#> s178              10.26284                0.5329996          8.944272
#>       h_Absorbance_Log h_Benzene_Log h_NO2_Log h_PM_Log h_TEX_Log
#> s548         0.2319273     0.8362600  6.172735 2.634545  3.079844
#> s555        -0.5441289     0.6195517  3.285681 2.011653  3.342032
#> s598        -0.7105256     0.4032863  2.678822 1.845124  4.283420
#> s395        -0.6001003     0.1976807  3.209300 2.190706  3.752039
#> s1135       -0.1702925     0.6871132  2.523871 2.702194  3.079750
#> s178        -0.8927670     0.8544012  3.132325 1.986648  3.160899
#>       e3_alcpreg_yn_None h_bfdur_Ter h_cereal_preg_Ter h_dairy_preg_Ter
#> s548                   1 (10.8,34.9]          (9,27.3]      (17.1,27.1]
#> s555                   0  (34.9,Inf]             (0,9]       (27.1,Inf]
#> s598                   1  (34.9,Inf]        (27.3,Inf]      (17.1,27.1]
#> s395                   0  (34.9,Inf]          (9,27.3]       (27.1,Inf]
#> s1135                  1 (10.8,34.9]             (0,9]      (17.1,27.1]
#> s178                   0    (0,10.8]        (27.3,Inf]      (17.1,27.1]
#>       h_fastfood_preg_Ter h_fish_preg_Ter h_folic_t1_None h_fruit_preg_Ter
#> s548           (0.83,Inf]       (1.9,4.1]               0       (0.6,18.2]
#> s555           (0.83,Inf]       (4.1,Inf]               0       (0.6,18.2]
#> s598          (0.25,0.83]       (1.9,4.1]               0       (0.6,18.2]
#> s395          (0.25,0.83]       (1.9,4.1]               0       (0.6,18.2]
#> s1135          (0.83,Inf]         (0,1.9]               1       (18.2,Inf]
#> s178           (0.83,Inf]         (0,1.9]               0       (18.2,Inf]
#>       h_legume_preg_Ter h_meat_preg_Ter h_pamod_t3_None h_pavig_t3_None
#> s548            (0.5,2]         (0,6.5]            None          Medium
#> s555            (2,Inf]        (6.5,10]      Very Often             Low
#> s598            (0,0.5]        (10,Inf]      Very Often             Low
#> s395            (0,0.5]        (10,Inf]           Often             Low
#> s1135           (0.5,2]        (10,Inf]      Very Often             Low
#> s178            (2,Inf]        (6.5,10]      Very Often          Medium
#>       h_veg_preg_Ter hs_bakery_prod_Ter hs_beverages_Ter hs_break_cer_Ter
#> s548      (8.8,16.5]              (2,6]          (1,Inf]        (5.5,Inf]
#> s555         (0,8.8]              (2,6]          (1,Inf]        (1.1,5.5]
#> s598      (16.5,Inf]              (2,6]          (1,Inf]        (1.1,5.5]
#> s395         (0,8.8]              (0,2]        (0.132,1]        (1.1,5.5]
#> s1135     (16.5,Inf]            (6,Inf]        (0,0.132]        (1.1,5.5]
#> s178         (0,8.8]            (6,Inf]          (1,Inf]          (0,1.1]
#>       hs_caff_drink_Ter hs_dairy_Ter hs_fastfood_Ter hs_KIDMED_None
#> s548          (0,0.132]  (14.6,25.6]       (0,0.132]              4
#> s555          (0,0.132]     (0,14.6]     (0.132,0.5]              2
#> s598        (0.132,Inf]     (0,14.6]       (0.5,Inf]              5
#> s395        (0.132,Inf]  (14.6,25.6]       (0.5,Inf]              7
#> s1135         (0,0.132]  (14.6,25.6]     (0.132,0.5]              4
#> s178        (0.132,Inf]   (25.6,Inf]     (0.132,0.5]              3
#>       hs_mvpa_prd_alt_None hs_org_food_Ter hs_pet_cat_r2_None
#> s548              19.71076       (0.132,1]                  0
#> s555              47.75195         (1,Inf]                  0
#> s598              31.83190         (1,Inf]                  0
#> s395             110.76032       (0.132,1]                  0
#> s1135             86.06060       (0,0.132]                  0
#> s178              12.12149         (1,Inf]                  0
#>       hs_pet_dog_r2_None hs_pet_None hs_proc_meat_Ter hs_readymade_Ter
#> s548                   0          No          (1.5,4]        (0.5,Inf]
#> s555                   0          No          (1.5,4]        (0.5,Inf]
#> s598                   0          No          (0,1.5]        (0.5,Inf]
#> s395                   0          No          (4,Inf]        (0.5,Inf]
#> s1135                  0          No          (1.5,4]      (0.132,0.5]
#> s178                   0          No          (1.5,4]        (0.5,Inf]
#>       hs_sd_wk_None hs_total_bread_Ter hs_total_cereal_Ter hs_total_fish_Ter
#> s548      183.57143           (7,17.5]         (14.1,23.6]           (3,Inf]
#> s555      570.00000              (0,7]            (0,14.1]           (1.5,3]
#> s598      166.42857              (0,7]            (0,14.1]           (3,Inf]
#> s395      117.85714           (7,17.5]         (14.1,23.6]           (1.5,3]
#> s1135     435.71429              (0,7]            (0,14.1]           (0,1.5]
#> s178       66.42857         (17.5,Inf]         (14.1,23.6]           (0,1.5]
#>       hs_total_fruits_Ter hs_total_lipids_Ter hs_total_meat_Ter
#> s548           (14.1,Inf]               (0,3]             (6,9]
#> s555                (0,7]               (3,7]             (6,9]
#> s598             (7,14.1]               (3,7]             (0,6]
#> s395           (14.1,Inf]               (3,7]           (9,Inf]
#> s1135            (7,14.1]               (0,3]             (0,6]
#> s178                (0,7]             (7,Inf]           (9,Inf]
#>       hs_total_potatoes_Ter hs_total_sweets_Ter hs_total_veg_Ter
#> s548                (4,Inf]             (0,4.1]        (8.5,Inf]
#> s555                (4,Inf]           (4.1,8.5]          (6,8.5]
#> s598                  (0,3]             (0,4.1]        (8.5,Inf]
#> s395                  (0,3]           (4.1,8.5]        (8.5,Inf]
#> s1135               (4,Inf]           (4.1,8.5]        (8.5,Inf]
#> s178                (4,Inf]           (8.5,Inf]        (8.5,Inf]
#>       hs_total_yog_Ter hs_dif_hours_total_None hs_as_c_Log2 hs_as_m_Log2
#> s548             (0,6]               10.277709    1.6553518    -3.929044
#> s555             (0,6]               10.371332   -5.4796130   -18.893931
#> s598             (0,6]                9.793991    0.2016339     3.000000
#> s395         (8.5,Inf]               11.182731    1.0285692     0.790772
#> s1135            (0,6]                9.622349   -2.7505391    -7.984624
#> s178         (8.5,Inf]               10.454553    1.2078929     1.286881
#>       hs_cd_c_Log2 hs_cd_m_Log2 hs_co_c_Log2 hs_co_m_Log2 hs_cs_c_Log2
#> s548     -5.993092   -2.9657843    -2.046921   -0.1568201   0.27500705
#> s555     -3.035047   -3.1328943    -4.025618   -2.1078033   0.08406426
#> s598     -3.738890   -3.4851960    -2.145605   -2.1140352   0.61353165
#> s395     -4.451662   -0.9351171    -2.634867   -2.9545570   0.82374936
#> s1135    -7.858989   -2.7466158    -1.831358   -1.1046974  -0.20257192
#> s178     -3.932361   -2.7466158    -3.095420    0.2265085   0.64154603
#>       hs_cs_m_Log2 hs_cu_c_Log2 hs_cu_m_Log2 hs_hg_c_Log2 hs_hg_m_Log2
#> s548    0.35614381     9.661778     9.890264    1.1243281  -1.27578631
#> s555    0.15055968     9.658211    10.958553   -1.2898273  -0.58424133
#> s598    1.01435529     9.640245    10.321928   -0.8943219   1.85199884
#> s395   -0.10159814     9.763212    10.321928   -0.1795147   0.57531233
#> s1135   0.02856915     9.658211    10.116344   -0.4150375   1.88981713
#> s178   -0.11872694     9.588715    10.216746    0.8953026  -0.06642736
#>       hs_mn_c_Log2 hs_mn_m_Log2 hs_mo_c_Log2 hs_mo_m_Log2 hs_pb_c_Log2
#> s548      2.912650     4.087463    1.2448871   -1.1976000     2.987321
#> s555      3.596935     3.536053    0.6040713   -1.7226103     2.891419
#> s598      2.929791     3.220330   -0.6060347   -0.1713684     2.568032
#> s395      3.001802     3.754888   -1.1392358   -0.7321646     2.232661
#> s1135     2.722466     2.629939    3.1715271    0.1110313     3.291309
#> s178      3.405992     3.666757   -0.5269924   -0.7321646     2.881665
#>       hs_pb_m_Log2 hs_tl_cdich_None hs_tl_mdich_None h_humidity_preg_None
#> s548      2.316146       Undetected       Undetected             82.30592
#> s555      3.030336       Undetected       Undetected             83.01580
#> s598      3.161888       Undetected       Undetected             69.73431
#> s395      2.575312       Undetected       Undetected             70.72809
#> s1135     3.195348       Undetected       Undetected             57.18486
#> s178      3.807355       Undetected       Undetected             74.95506
#>       h_pressure_preg_None h_temperature_preg_None hs_hum_mt_hs_h_None
#> s548              980.8294                7.982746            68.20094
#> s555              980.8445                9.808083            82.44498
#> s598             1000.8884               13.698591            85.46740
#> s395             1000.2018                9.876204            89.29032
#> s1135            1015.3033               21.419140            57.19452
#> s178             1003.0392               12.886891            71.39742
#>       hs_tm_mt_hs_h_None hs_uvdvf_mt_hs_h_None hs_hum_dy_hs_h_None
#> s548            6.632258             0.2676667            68.91667
#> s555           13.717742             1.3237931            80.09360
#> s598            4.058379             0.0330000            59.00000
#> s395            6.519355             0.3062069            99.00000
#> s1135          25.974194             5.1133333            69.99000
#> s178           16.128752             2.9860000            83.09333
#>       hs_hum_wk_hs_h_None hs_tm_dy_hs_h_None hs_tm_wk_hs_h_None
#> s548             59.88021           6.529167           6.041406
#> s555             80.23743           7.133320          13.367036
#> s598             91.12500          -2.900000          -1.025000
#> s395             98.25000           2.600000           4.450000
#> s1135            64.11750          25.600000          25.475000
#> s178             67.23368          20.383871          15.925000
#>       hs_uvdvf_dy_hs_h_None hs_uvdvf_wk_hs_h_None hs_blueyn300_s_None
#> s548                   0.28            0.28714286                   0
#> s555                   1.38            1.75142857                   0
#> s598                   0.01            0.01428571                   1
#> s395                   0.12            0.17428571                   0
#> s1135                  5.35            4.87428571                   0
#> s178                   3.37            3.31857143                   0
#>       h_blueyn300_preg_None h_greenyn300_preg_None h_ndvi100_preg_None
#> s548                      0                      1           0.3309343
#> s555                      0                      1           0.3897694
#> s598                      1                      0           0.4118056
#> s395                      0                      1           0.6900000
#> s1135                     0                      1           0.1373250
#> s178                      0                      1           0.5318743
#>       hs_greenyn300_s_None hs_blueyn300_h_None hs_greenyn300_h_None
#> s548                     1                   0                    1
#> s555                     1                   0                    1
#> s598                     1                   1                    1
#> s395                     1                   0                    1
#> s1135                    1                   0                    0
#> s178                     0                   0                    1
#>       hs_ndvi100_h_None hs_ndvi100_s_None h_lden_cat_preg_None hs_ln_cat_h_None
#> s548          0.2619719         0.3700306             65.38800                4
#> s555          0.4859200         0.5260056             39.41497                2
#> s598          0.8076286         0.6063735             65.38800                2
#> s395          0.4668250         0.5365706             50.00000                2
#> s1135         0.5183771         0.3507906             71.29236                1
#> s178          0.6309914         0.4598028             64.36400                2
#>       hs_lden_cat_s_None hs_dde_cadj_Log2 hs_dde_madj_Log2 hs_ddt_cadj_Log2
#> s548                   3         4.397398         7.280566        1.4965645
#> s555                   2         5.073660         3.967523        0.2209479
#> s598                   1         3.184601         3.743854       -1.4298728
#> s395                   2         5.021621         4.993255      -10.2899871
#> s1135                  3         6.058512         8.507109        0.8078187
#> s178                   1         3.228497         5.786281       -4.0315477
#>       hs_ddt_madj_Log2 hs_hcb_cadj_Log2 hs_hcb_madj_Log2 hs_pcb118_cadj_Log2
#> s548         0.8023278         2.328545         4.058057           0.7261138
#> s555         0.6764266         4.603839         2.034068           1.8140500
#> s598        -1.0738303         3.374735         2.676388           1.6977361
#> s395        -0.5898499         3.908736         2.958089           2.2821563
#> s1135        1.8622269         3.250550         3.346906           0.9371018
#> s178        -0.5401198         2.093691         2.925833           1.2004458
#>       hs_pcb118_madj_Log2 hs_pcb138_cadj_Log2 hs_pcb138_madj_Log2
#> s548            0.8773249            2.103670            3.450788
#> s555            1.0521823            3.602497            1.787701
#> s598            1.5415867            3.121447            2.937510
#> s395            2.0683998            4.141065            3.768623
#> s1135           0.2991493            1.923029            2.178092
#> s178            1.8856717            1.838072            3.998914
#>       hs_pcb153_cadj_Log2 hs_pcb153_madj_Log2 hs_pcb170_cadj_Log2
#> s548             3.427075            4.398305           0.6506521
#> s555             4.474462            2.852110           1.8193741
#> s598             4.291712            4.057322           1.7768773
#> s395             5.189689            4.677404           2.5954040
#> s1135            3.046468            3.270783          -0.1944951
#> s178             2.850649            5.343338           1.9342381
#>       hs_pcb170_madj_Log2 hs_pcb180_cadj_Log2 hs_pcb180_madj_Log2
#> s548            0.4639773           1.5324170            2.840244
#> s555            0.4639773           3.1056819            2.427232
#> s598            1.9412515           3.2731852            3.653995
#> s395            2.0650106           4.1424622            3.820939
#> s1135           1.3968458           1.0001747            2.477865
#> s178            3.6451235           0.6216219            4.977054
#>       hs_sumPCBs5_cadj_Log2 hs_sumPCBs5_madj_Log2 hs_dep_cadj_Log2
#> s548               4.529040              3.203309        -5.621780
#> s555               5.640019              3.203309        -3.874674
#> s598               5.479832              6.531905         2.643732
#> s395               6.692561              5.981735         2.093970
#> s1135              4.074065              6.084188         3.335985
#> s178               4.612107              6.772989        -6.073264
#>       hs_dep_madj_Log2 hs_detp_cadj_Log2 hs_detp_madj_Log2 hs_dmdtp_cdich_None
#> s548         3.3189632        -0.1968193         3.0292014          Undetected
#> s555         1.5457521        -2.8543646         1.0078850          Undetected
#> s598        -0.6129715        -5.0395139        -6.5515954          Undetected
#> s395         2.1288965         0.6179130         1.6973774          Undetected
#> s1135        2.1345302         3.2053664         0.2568657          Undetected
#> s178         1.8602252         0.8498652        -7.8249558          Undetected
#>       hs_dmp_cadj_Log2 hs_dmp_madj_Log2 hs_dmtp_cadj_Log2 hs_dmtp_madj_Log2
#> s548        -3.9134270         5.772154         0.3202063          6.642612
#> s555         0.3406811         2.194291        -2.2143799          1.200211
#> s598         0.4393773         2.194291         1.8175012          1.438817
#> s395        -7.1109560         5.212931         1.6536180          5.496598
#> s1135        2.5499939         4.140999         1.9674973          5.523995
#> s178         1.0094394         2.599798         2.9916697          2.441006
#>       hs_pbde153_cadj_Log2 hs_pbde153_madj_Log2 hs_pbde47_cadj_Log2
#> s548            -0.2220472          -2.02195748           -1.327538
#> s555            -8.2312357           0.06158347           -2.449476
#> s598            -1.9043665         -10.85822494           -2.365715
#> s395            -2.2771770          -1.45803551           -1.986786
#> s1135          -13.0227177           0.06158347          -11.190074
#> s178           -11.4125543          -0.72159095           -2.214787
#>       hs_pbde47_madj_Log2 hs_pfhxs_c_Log2 hs_pfhxs_m_Log2 hs_pfna_c_Log2
#> s548           -1.8333971      -1.4833867      -0.2529789     -0.4205499
#> s555           -0.2313590      -2.4041523      -1.7230151     -0.8139326
#> s598           -0.7402796      -1.1239863      -1.4477809      0.9039864
#> s395           -0.6862967      -0.3054900      -1.3637356     -1.0512585
#> s1135          -0.2313590      -3.2957240      -1.7705798     -0.8934231
#> s178           -0.8263363      -0.4557693      -0.2330125     -0.5361877
#>       hs_pfna_m_Log2 hs_pfoa_c_Log2 hs_pfoa_m_Log2 hs_pfos_c_Log2
#> s548     -0.32565470    0.575337219    1.837058207    -0.02184093
#> s555     -0.05451747    0.789403199    0.008883347     0.31034727
#> s598     -5.74758539    1.043480466    0.147029348     1.16281873
#> s395     -0.70319902    0.897028002    1.605374581     1.77472838
#> s1135     0.41086402    0.766020855    1.568974543     0.48611542
#> s178     -0.94370886    0.006974908    1.914157494     1.66382171
#>       hs_pfos_m_Log2 hs_pfunda_c_Log2 hs_pfunda_m_Log2 hs_bpa_cadj_Log2
#> s548        1.973866        -2.859890        -3.212220        2.7934139
#> s555        2.363030        -7.581327        -2.612664        4.1444726
#> s598        2.087922        -4.285316        -1.818424        1.4320657
#> s395        3.160208        -3.865110        -2.413287        1.5370346
#> s1135       1.886701        -4.997875        -2.915871        3.9929597
#> s178        3.225800        -4.624617        -3.490419        0.8748443
#>       hs_bpa_madj_Log2 hs_bupa_cadj_Log2 hs_bupa_madj_Log2 hs_etpa_cadj_Log2
#> s548        1.16763084        -3.2085230          6.742633        -1.1156996
#> s555        0.05551944        -0.2497832         -1.237360         3.9009840
#> s598        5.14723226        -5.7183256          6.333751        -0.9372423
#> s395        2.84965889        -4.2118856          1.592276        -1.1699494
#> s1135       5.02859543        -4.2759953         -0.910728         3.3071302
#> s178        0.83178137        -9.8253317         -2.308190        -0.8220723
#>       hs_etpa_madj_Log2 hs_mepa_cadj_Log2 hs_mepa_madj_Log2 hs_oxbe_cadj_Log2
#> s548          5.4839829       5.530419435          8.018891         2.1065968
#> s555          4.0653328       7.155088119          6.972590        -0.5965608
#> s598          5.7990156      -0.006212212          8.133136         0.9232409
#> s395         12.7256918       3.990299542          6.972590         1.7788307
#> s1135         2.1128397       5.275220421          4.601472         3.4310270
#> s178         -0.2066972       1.996507969          5.290413        -1.7710243
#>       hs_oxbe_madj_Log2 hs_prpa_cadj_Log2 hs_prpa_madj_Log2 hs_trcs_cadj_Log2
#> s548           2.439743          2.929369          6.099105        -1.0922651
#> s555          -1.520746         -4.830559          5.774825         4.0059123
#> s598           8.671510         -4.761428          9.634782         0.2138049
#> s395           6.274219         -7.337670          6.742139        -2.1521805
#> s1135          5.487510          2.380210          1.575864         0.3766612
#> s178           6.580407         -4.553921          2.909095        -1.9094638
#>       hs_trcs_madj_Log2 hs_mbzp_cadj_Log2 hs_mbzp_madj_Log2 hs_mecpp_cadj_Log2
#> s548           6.611306          3.356539          1.791004           5.077850
#> s555           2.311827          3.818532          2.991920           6.360956
#> s598           9.566305          2.532486          2.198102           5.118947
#> s395           6.308943          1.690778          1.120067           4.370613
#> s1135          8.443111          4.304188          3.363802           6.570902
#> s178           9.337416          1.963026          3.627870           4.169406
#>       hs_mecpp_madj_Log2 hs_mehhp_cadj_Log2 hs_mehhp_madj_Log2
#> s548            4.262406           4.479090           2.966073
#> s555            4.326828           5.049778           3.709055
#> s598            5.521025           3.942982           3.726604
#> s395            4.682827           2.408466           3.797013
#> s1135           5.386673           4.944154           4.584204
#> s178            5.333671           3.358253           4.646188
#>       hs_mehp_cadj_Log2 hs_mehp_madj_Log2 hs_meohp_cadj_Log2 hs_meohp_madj_Log2
#> s548          2.3883178          1.984630           4.526982           2.733261
#> s555          2.3136518          3.808210           4.882229           4.360200
#> s598          1.8776791          3.888287           3.499562           3.384968
#> s395          0.3708812          3.385743           1.702568           3.018504
#> s1135         2.5552289          2.340926           4.936904           3.927649
#> s178          0.6779291          3.274544           2.785315           4.110425
#>       hs_mep_cadj_Log2 hs_mep_madj_Log2 hs_mibp_cadj_Log2 hs_mibp_madj_Log2
#> s548          6.956249         8.479863          4.007595          4.592135
#> s555          4.725078         8.713375          7.004560          4.976039
#> s598          5.219245        10.690529          4.443710          5.978889
#> s395          2.305678         6.543938          3.612705          3.762709
#> s1135         5.911079        11.460176          5.807431          5.533377
#> s178          6.765690         6.261872          4.925551          4.686872
#>       hs_mnbp_cadj_Log2 hs_mnbp_madj_Log2 hs_ohminp_cadj_Log2
#> s548           4.096990          3.799163            2.337238
#> s555           5.766602          5.357187            3.740723
#> s598           5.052293          5.405388            1.908719
#> s395           3.372310          3.102331            1.033223
#> s1135          6.311060          4.938192            2.897308
#> s178           4.882002          5.981519            2.954178
#>       hs_ohminp_madj_Log2 hs_oxominp_cadj_Log2 hs_oxominp_madj_Log2
#> s548           -0.4079240           2.59419576           -0.4436485
#> s555           -0.5650174           3.15568465           -0.6939739
#> s598            0.7619635           1.43555340            0.3585584
#> s395            3.1362128          -0.03892898            1.7285519
#> s1135          -0.4683972           2.42987842           -0.7125105
#> s178           -0.5542561           1.87135415           -0.0589566
#>       hs_sumDEHP_cadj_Log2 hs_sumDEHP_madj_Log2 FAS_cat_None
#> s548              6.399445             5.154776         High
#> s555              6.970544             5.805357         High
#> s598              6.575202             6.298866       Middle
#> s395              4.340647             5.768232         High
#> s1135             8.030191             6.270853       Middle
#> s178              5.002653             6.429820         High
#>       hs_contactfam_3cat_num_None hs_hm_pers_None hs_participation_3cat_None
#> s548                  Once a week               4                       None
#> s555               (almost) Daily               4                       None
#> s598               (almost) Daily               5    2 or more organisations
#> s395                  Once a week               4    2 or more organisations
#> s1135              (almost) Daily               4                       None
#> s178               (almost) Daily               4                       None
#>       e3_asmokcigd_p_None hs_cotinine_cdich_None hs_cotinine_mcat_None
#> s548            0.0000000             Undetected           Non-smokers
#> s555            0.0000000             Undetected               Smokers
#> s598            0.0000000             Undetected           Non-smokers
#> s395            0.0000000             Undetected           Non-smokers
#> s1135           0.6666666             Undetected           SHS smokers
#> s178            0.0000000             Undetected           SHS smokers
#>       hs_globalexp2_None hs_smk_parents_None h_distinvnear1_preg_Log
#> s548            exposure             neither               -4.148408
#> s555         no exposure             neither               -4.740835
#> s598         no exposure             neither               -5.718191
#> s395         no exposure             neither               -2.841497
#> s1135           exposure                both                0.501706
#> s178         no exposure             neither               -6.181876
#>       h_trafload_preg_pow1over3 h_trafnear_preg_pow1over3
#> s548                 68.6958953                  7.937005
#> s555                  0.3458301                 23.990158
#> s598                113.0812135                  7.937005
#> s395                  0.3458301                  7.937005
#> s1135               139.5134895                 10.811445
#> s178                  0.3458301                 10.000000
#>       hs_trafload_h_pow1over3 hs_trafnear_h_pow1over3 h_bro_preg_Log
#> s548                135.99576               44.968542      2.6710404
#> s555                139.29800               26.711608     -1.6044503
#> s598                168.02106                7.937005     -0.5430045
#> s395                 56.66268                7.937005     -0.7318880
#> s1135                65.06702                7.937005      2.6710404
#> s178                114.86848               10.000000      2.0083483
#>       h_clf_preg_Log h_thm_preg_Log h_cohort e3_sex_None e3_yearbir_None
#> s548       3.6368129      3.8261620        3      female            2006
#> s555      -0.4758875      0.4719939        4        male            2008
#> s598       2.7726512      2.8173143        5        male            2005
#> s395       1.7919261      1.8918844        5        male            2007
#> s1135     -4.4191264      1.5373435        6        male            2008
#> s178      -6.9077553      2.0283922        2      female            2004
#>       h_mbmi_None hs_wgtgain_None e3_gac_None h_age_None h_edumc_None
#> s548     18.90690              NA    40.85714   35.73717            2
#> s555     30.10212              14    39.00000   24.61187            2
#> s598     20.46334              28    39.00000   36.00000            3
#> s395     22.14533               7    41.14286   35.07574            3
#> s1135    19.13265              19    38.57143   24.58607            2
#> s178     19.33373              21    40.14286   32.25462            3
#>       h_native_None h_parity_None hs_child_age_None hs_c_height_None
#> s548              2             0          8.421743            1.325
#> s555              2             0          6.102669            1.240
#> s598              2             2          8.917180            1.405
#> s395              2             1          6.921173            1.250
#> s1135             2             1          6.620009            1.250
#> s178              2             2         10.699521            1.455
#>       hs_c_weight_None e3_bw hs_asthma hs_zbmi_who hs_correct_raven hs_Gen_Tot
#> s548              30.1  3530         0        0.65               30    19.2500
#> s555              21.0  3370         0       -1.40               22    48.0000
#> s598              35.9  4328         0        1.15               33     8.0625
#> s395              28.2  3242         0        0.47               24     4.0000
#> s1135             22.7  3300         0       -0.68               23    12.0625
#> s178              26.1  3215         0        0.62               32    22.1875
#>       hs_bmi_c_cat
#> s548             2
#> s555             2
#> s598             3
#> s395             2
#> s1135            2
#> s178             2