Multi-Omics Assays for the Exposome Dataset

A named list of omics data matrices with features × samples format. Each assay contains processed data from a different platform.

Usage

omics_list

Format

A named list with the following components:

Gene Expression

A matrix or data.frame of gene expression values.

Methylation

A matrix or data.frame of methylation levels. Probes have been filtered to only include the top 20% most variant probes.

Metabolomics

A matrix or data.frame of metabolite concentrations.

Source

Derived from the ISGlobal Exposome Data Challenge 2021.

Examples

data(omics_list)
names(omics_list)
#> [1] "Gene Expression" "Methylation"     "Metabolomics"   
lapply(omics_list, dim)
#> $`Gene Expression`
#> [1] 28738  1007
#> 
#> $Methylation
#> [1] 59844   918
#> 
#> $Metabolomics
#> [1]  177 1198
#>