Plot Exposures Associated with Features in GO Groups

Visualizes exposures that are significantly associated with features in specified Gene Ontology (GO) groups.

Usage

plot_go_group_exposures(
  expomicset,
  go_groups,
  geneset = "deg_exp_cor",
  feature_col = "feature",
  top_n = 15
)

Arguments

expomicset

A MultiAssayExperiment object containing functional enrichment and exposure-feature correlation results.

go_groups

A character vector specifying GO groups of interest. Use "all" to include all GO groups.

geneset

A character string indicating the gene set to use. Options are "deg_exp_cor" for DEGs or "factor_exp_cor" for factors. Default is "deg_exp_cor".

feature_col

A character string specifying the feature column to match with correlated features. Default is "feature".

top_n

An integer specifying the number of top exposures to display per experimental assay. Default is 15.

Value

A ggplot object displaying the top exposures associated with features in GO groups.

Details

This function identifies exposures significantly correlated with features within specified GO groups. It extracts functional enrichment results from metadata(expomicset)$functional_enrichment and cross-references them with exposure-feature correlation results:

"deg_exp_cor": Uses metadata(expomicset)$omics_exposure_deg_correlation

"factor_exp_cor": Uses metadata(expomicset)$omics_exposure_factor_correlation

The resulting plot:

Displays exposures grouped by experimental assay (exp_name).

Uses bar heights to indicate the number of associated features.

Colors bars based on exposure categories.

Examples

if (FALSE) { # \dontrun{
plot_go_group_exposures(expom, go_groups = c("GO:0006355", "GO:0045893"))
} # }